# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Hausa Live Speech-to-Text API
  description: |
    A comprehensive API for Hausa speech-to-text transcription with subscription tiers, 
    points system, and premium features. Supports both file upload and live streaming transcription.
  version: 0.1.0
  contact:
    name: Hausa Speech API Support
    email: support@hausaspeech.com
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.hausaspeech.com
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/auth/register:
    post:
      summary: Register new user
      description: |
        Register a new user with email and password. Creates user profile in MongoDB
        with default free tier subscription. Password is hashed using bcrypt.
      operationId: registerUser
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: User's password (minimum 8 characters)
                name:
                  type: string
                  description: User's display name
                preferences:
                  $ref: '#/components/schemas/UserPreferences'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/AuthUser'
                  token:
                    type: string
                    description: JWT token for authentication
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/points/balance:
    get:
      summary: Get points balance
      description: Get current user's points balance and limits
      operationId: getPointsBalance
      tags: [Points]
      responses:
        '200':
          description: Points balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  points:
                    type: integer
                  tier:
                    type: string
                  limits:
                    $ref: '#/components/schemas/PointsLimits'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/points/watch-ad:
    post:
      summary: Watch rewarded ad
      description: Record ad watch and award points to user
      operationId: watchAd
      tags: [Points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [adId]
              properties:
                adId:
                  type: string
                  description: Unique identifier for the ad
      responses:
        '200':
          description: Points awarded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pointsEarned:
                    type: integer
                  newBalance:
                    type: integer
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/points/summarize:
    post:
      summary: Generate summary using points
      description: Generate a summary for short transcripts using points
      operationId: summarizeWithPoints
      tags: [Points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                transcriptId:
                  type: string
                content:
                  type: string
                  description: Transcript content to summarize
                duration:
                  type: number
                  description: Duration in seconds (max 60 for short summary)
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  summary:
                    type: string
                  pointsSpent:
                    type: integer
                  remainingPoints:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientPoints'
  /api/points/punctuation:
    post:
      summary: Fix punctuation using points
      description: Fix punctuation and grammar for one paragraph using points
      operationId: fixPunctuationWithPoints
      tags: [Points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: Text to fix punctuation (max 500 characters)
      responses:
        '200':
          description: Punctuation fixed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  originalContent:
                    type: string
                  fixedContent:
                    type: string
                  pointsSpent:
                    type: integer
                  remainingPoints:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientPoints'
  /api/points/translate:
    post:
      summary: Translate content using points
      description: Translate short text using points
      operationId: translateWithPoints
      tags: [Points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, targetLanguage]
              properties:
                content:
                  type: string
                  description: Text to translate
                targetLanguage:
                  type: string
                  enum: [ha-NG, en-US, fr-FR, ar]
                  description: Target language for translation
                duration:
                  type: number
                  description: Duration in seconds (max 60 for short translation)
      responses:
        '200':
          description: Translation completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  originalContent:
                    type: string
                  translatedContent:
                    type: string
                  targetLanguage:
                    type: string
                  pointsSpent:
                    type: integer
                  remainingPoints:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/InsufficientPoints'

  # Transcription Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Token for MongoDB authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for service-to-service authentication
  schemas:
    AuthUser:
      type: object
      properties:
        id:
          type: string
          description: Unique user ID
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's display name
        subscriptionTier:
          type: string
          enum: [free, basic, gold, premium]
          description: Current subscription tier
        subscriptionStatus:
          type: string
          enum: [active, cancelled, expired, trial]
          description: Subscription status
        pointsBalance:
          type: integer
          description: Current points balance
        isPremium:
          type: boolean
          description: Whether user has premium features
        usageStats:
          $ref: '#/components/schemas/UsageStats'
    UsageStats:
      type: object
      properties:
        dailyMinutes:
          type: integer
          description: Minutes used today
        monthlyMinutes:
          type: integer
          description: Minutes used this month
        totalMinutes:
          type: integer
          description: Total minutes used
        transcriptsCount:
          type: integer
          description: Total transcripts created
    UserPreferences:
      type: object
      properties:
        language:
          type: string
          enum: [ha-NG, en-US, fr-FR, ar]
          description: Preferred language
        theme:
          type: string
          enum: [light, dark]
          description: UI theme preference
        autoPunctuation:
          type: boolean
          description: Enable automatic punctuation
        cloudSync:
          type: boolean
          description: Enable cloud synchronization
...
    PointsLimits:
      type: object
      properties:
        dailyAdWatches:
          type: integer
          description: Maximum ads per day
        maxPointsBalance:
          type: integer
          description: Maximum points balance
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
              details:
                type: object
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...
    InsufficientPoints:
      description: Insufficient points
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
              details:
                type: object
                properties:
                  actionId:
                    type: string
                  cost:
                    type: integer
                  remainingPoints:
                    type: integer
                  reason:
                    type: string
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              code:
                type: string
...